refactor: modularize NFT gallery into components and composables
